Handling disconnections better
In the initial implementation, it was impossible to determine if a
Watcher was disconnected from any possible updates. These changes add
Result returns in various locations that can be used to detect the
difference between timeouts, new values being available, or
disconnections.

Cloning is now fully implemented and tested, proving mpmc.