Fixed race condition with updating values
Apparently the stress test was too stressful to uncover the particular
race condition I discovered in Sediment. Because I was often only
updating a Watchable once per database open, I uncovered a race
condition when that single update wouldn't be read in some situations.

This ultimately boiled down to the logic in create_listener_if_needed. I
needed a second check to ensure the value was still considered current
after creating the listener, otherwise that listener would never wake up
for the value that was just written in another thread while the listener
was being created.

After fixing this, however, I found that the stress test unit test was
breaking. I was experiencing spurious wake-ups from waiting for an
event. So each place I was calling create_listener_if_needed, I needed
to loop and re-call if the event listener woke up but we didn't detect a
change. Previously, this only happened when a disconnect or a new value
was written. Now, disconnects are always handled by
create_listener_if_needed.