Fixed Watcher::clone reqs, made Watcher atomic
While working on a crate that combines this and the notify crate to
provide a simple file watching abstraction, I ran into a limitation
where a StaticFileWatcher was needing to provide a different API due to
some APIs requiring mutating the Watcher.

I realized that the version could be swapped to an AtomicUsize for
negligable impact on performance, and the benefit would be that Watcher
wouldn't require mut anymore.